More code cleanup.  Corrected the fire order of events when created the ClustedCacheFactory.
in clusterMasterWatcher the method: public void viewAccepted(View v)

modified this so that the current node is listed in the nodes array.  This facilitates the plugin starting with only 1 node in the cluster, which is ok, we shouldn't need to wait for additional nodes before starting the plugin (or the server).  Additional nodes will be added to the cluster group as they come online.