Temporary fix for reading privacy lists in ClusteredClientSession.  When these lists are null from the ObjectInput stream an exception is thrown and some messages are never routed.  If the default constructor is used again a exception is thrown and messages are never routed.
As a temporary fix I have set these lists to null which is the default if the user has not yet created a list.  This hack is just so I can work on and test other portions of the clustering plugin.

With this commit most of the plugin is functional.  There are still defects related to nodes joining and leaving the cluster (users are unable to log in if a node is dropped).  When a node is dropped from the cluster user persistence from the dropped node is not shown as offline to users still connected to the cluster.  As well as other things I'm sure I haven't found yet.